docs: move docs tooling to _tooling
Issue-ref: see #139
