+++ docs/_tooling/assets/css/score_design.css
