+++ docs/_tooling/incremental.py
@@ -12,31 +12,23 @@
 # *******************************************************************************
 
 import os
-import sys
 
-try:
-    from sphinx.cmd.build import main as sphinx_main
-except ImportError:
-    sys.exit(
-        "This script must be run from Bazel via `bazel run //docs:incremental`"
-        " (sphinx not installed)."
-    )
-
-workspace = os.getenv("BUILD_WORKSPACE_DIRECTORY")
-if not workspace:
-    sys.exit(
-        "This script must be run from Bazel via `bazel run //docs:incremental`"
-        " (BUILD_WORKSPACE_DIRECTORY not set)."
-    )
+from sphinx.cmd.build import main as sphinx_main
 
 # sphinx will print relative paths to the current directory.
 # Change to the workspace root so that the paths are readable and clickable.
-os.chdir(workspace)
+workspace = os.getenv("BUILD_WORKSPACE_DIRECTORY")
+if workspace:
+    os.chdir(workspace)
+
+
 sphinx_main(
     [
-        "docs",
-        "_build",
+        "docs",  # src dir
+        "_build",  # out dir
         "--jobs",
         "auto",
+        "--conf-dir",
+        "docs",
     ]
 )
